requirements.txt
# Flask and Extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter for production

# Forms and Validation
WTForms==3.1.1
email-validator==2.1.0

# Authentication and Security
Werkzeug==3.0.1
python-dotenv==1.0.0

# WSGI Server
gunicorn==21.2.0

# SEO and Utilities
python-slugify==8.0.1
Pillow==10.1.0  # Image processing

# Date and Time
python-dateutil==2.8.2
